fix(frame): change validSig from uint32_t to uint16_t
Changed signature field from uint32_t (bytes 0-3) to uint16_t (bytes 0-1)
to keep all other metadata fields at their original offsets.

Frame layout now:
- Bytes 0-1: validSig (uint16_t, value 0xBA77)
- Byte 2: version (uint8_t)
- Byte 3: cellBufferStart (uint8_t)
- Bytes 4-5: frameBytes (uint16_t)
- All other fields maintain original offsets

This prevents breaking changes to existing offset calculations.

+++ STORE.h
@@ -53,7 +53,7 @@ typedef struct
 
 // Frame metadata structure - contains all the control and status fields
 typedef struct __attribute__((aligned(4))) {
-	uint32_t validSig;  // if != FRAME_VALID_SIG then frame is not valid (was 0xBA77DA7A, now 0x0000BA77)
+	uint16_t validSig;  // if != FRAME_VALID_SIG then frame is not valid (0xBA77)
 	uint8_t version;    // Frame format version (for future compatibility)
 	uint8_t cellBufferStart;  // Offset from start of frame to cell data buffer (4-byte aligned, max 255)
 	uint16_t frameBytes;  // actual number of bytes per frame, for diagnostics
